Fixing for larger word list

+++ constants.hh
@@ -7,6 +7,6 @@ constexpr size_t DIM = 9;
 // Frequently we index based on length, so include one extra (since index 0 isn't used)
 constexpr size_t SIZE = DIM + 1;
 
-using WordIndex = uint16_t;
+using WordIndex = uint32_t;
 
 constexpr size_t NUM_THREADS = 6;
